ARM64 jit: Fix some pointer truncations and other fun stuff..

+++ Core/MIPS/JitCommon/JitCommon.cpp
@@ -90,7 +90,7 @@ std::vector<std::string> DisassembleArm64(const u8 *data, int size) {
 	for (int i = 0; i < size; i += 4) {
 		const u32 *codePtr = (const u32 *)(data + i);
 		u32 inst = codePtr[0];
-		Arm64Dis((u32)(intptr_t)codePtr, inst, temp, sizeof(temp), false);
+		Arm64Dis((intptr_t)codePtr, inst, temp, sizeof(temp), false);
 		std::string buf = temp;
 		if (buf == "BKPT 1") {
 			bkpt_count++;
